backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
flasgger==0.9.7.1
PyYAML==6.0.2